+++ src/main/kotlin/mobi/hsz/idea/gitignore/ui/GeneratorDialog.kt
@@ -41,6 +41,9 @@ import mobi.hsz.idea.gitignore.ui.template.TemplateTreeNode
 import mobi.hsz.idea.gitignore.ui.template.TemplateTreeRenderer
 import mobi.hsz.idea.gitignore.util.Constants
 import mobi.hsz.idea.gitignore.util.Resources
+import mobi.hsz.idea.gitignore.util.Resources.Template.Container
+import mobi.hsz.idea.gitignore.util.Resources.Template.Container.STARRED
+import mobi.hsz.idea.gitignore.util.Resources.Template.Container.USER
 import mobi.hsz.idea.gitignore.util.Utils.createPreviewEditor
 import mobi.hsz.idea.gitignore.util.Utils.getWords
 import org.jetbrains.annotations.NonNls
@@ -81,23 +84,23 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
     private val root = TemplateTreeNode()
 
     /** Templates tree with checkbox feature. */
-    private var tree: CheckboxTree? = null
+    private lateinit var tree: CheckboxTree
 
     /** Tree expander responsible for expanding and collapsing tree structure. */
     private var treeExpander: DefaultTreeExpander? = null
 
     /** Dynamic templates filter. */
-    private var profileFilter: FilterComponent? = null
-
-    /** Preview editor with syntax highlight. */
-    private var preview: Editor? = null
+    private var profileFilter = TemplatesFilterComponent()
 
     /** [Document] related to the [Editor] feature. */
-    private var previewDocument: Document? = null
+    private val previewDocument = EditorFactory.getInstance().createDocument("")
+
+    /** Preview editor with syntax highlight. */
+    private val preview = createPreviewEditor(previewDocument, project, true)
 
     /** CheckboxTree selection listener. */
     private val treeSelectionListener = TreeSelectionListener { _: TreeSelectionEvent? ->
-        tree?.selectionPath?.let { updateDescriptionPanel(it) }
+        tree.selectionPath?.let { updateDescriptionPanel(it) }
     }
 
     companion object {
@@ -112,18 +115,9 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
          * @param container container type to search
          * @return group node
          */
-        private fun getGroupNode(root: TemplateTreeNode, container: Resources.Template.Container): TemplateTreeNode {
-            val childCount = root.childCount
-            (0 until childCount).forEach {
-                val child = root.getChildAt(it) as TemplateTreeNode
-                if (container == child.container) {
-                    return child
-                }
-            }
-            return TemplateTreeNode(container).apply {
-                root.add(this)
-            }
-        }
+        private fun getGroupNode(root: TemplateTreeNode, container: Container) =
+            root.children().toList().find { it is TemplateTreeNode && it.container == container } as TemplateTreeNode?
+                ?: TemplateTreeNode(container).apply { root.add(this) }
     }
 
     init {
@@ -136,17 +130,9 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
     override fun getPreferredFocusedComponent() = profileFilter
 
     override fun dispose() {
-        tree!!.removeTreeSelectionListener(treeSelectionListener)
-        EditorFactory.getInstance().releaseEditor(preview!!)
+        tree.removeTreeSelectionListener(treeSelectionListener)
+        EditorFactory.getInstance().releaseEditor(preview)
         super.dispose()
-    }
-
-    override fun show() {
-        if (ApplicationManager.getApplication().isUnitTestMode) {
-            dispose()
-            return
-        }
-        super.show()
     }
 
     override fun doOKAction() {
@@ -166,12 +152,11 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
 
         val iterator: Iterator<Resources.Template?> = checked.iterator()
         while (iterator.hasNext()) {
-            val template = iterator.next()
-            if (template != null) {
+            iterator.next()?.let {
                 content
-                    .append(message("file.templateSection", template.name))
+                    .append(message("file.templateSection", it.name))
                     .append(Constants.NEWLINE)
-                    .append(template.content)
+                    .append(it.content)
 
                 if (iterator.hasNext()) {
                     content.append(Constants.NEWLINE)
@@ -202,41 +187,42 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
         myOKAction = OptionOkAction()
     }
 
-    override fun createCenterPanel(): JComponent {
-        // general panel
-        val centerPanel = JPanel(BorderLayout())
-        centerPanel.preferredSize = Dimension(800, 500)
+    override fun createCenterPanel() = JPanel(BorderLayout()).apply {
+        preferredSize = Dimension(800, 500)
 
         // splitter panel - contains tree panel and preview component
-        val splitter = JBSplitter(false, 0.4f)
-        centerPanel.add(splitter, BorderLayout.CENTER)
-        val treePanel = JPanel(BorderLayout())
-        previewDocument = EditorFactory.getInstance().createDocument("")
-        preview = createPreviewEditor(previewDocument!!, project, true)
-        splitter.firstComponent = treePanel
-        splitter.secondComponent = preview!!.component
+        add(
+            JBSplitter(false, 0.4f).apply {
+                firstComponent = JPanel(BorderLayout()).apply {
+                    val treeScrollPanel = createTreeScrollPanel()
 
-        /* Scroll panel for the templates tree. */
-        val treeScrollPanel = createTreeScrollPanel()
-        treePanel.add(treeScrollPanel, BorderLayout.CENTER)
-        val northPanel = JPanel(GridBagLayout())
-        northPanel.border = JBUI.Borders.empty(2, 0)
-        northPanel.add(
-            createTreeActionsToolbarPanel(treeScrollPanel).component,
-            GridBagConstraints(
-                0, 0, 1, 1, 1.0, 1.0, GridBagConstraints.BASELINE_LEADING,
-                GridBagConstraints.HORIZONTAL, JBUI.emptyInsets(), 0, 0
-            )
+                    /* Scroll panel for the templates tree. */
+                    add(treeScrollPanel, BorderLayout.CENTER)
+                    add(
+                        JPanel(GridBagLayout()).apply {
+                            border = JBUI.Borders.empty(2, 0)
+                            add(
+                                createTreeActionsToolbarPanel(treeScrollPanel).component,
+                                GridBagConstraints(
+                                    0, 0, 1, 1, 1.0, 1.0, GridBagConstraints.BASELINE_LEADING,
+                                    GridBagConstraints.HORIZONTAL, JBUI.emptyInsets(), 0, 0
+                                )
+                            )
+                            add(
+                                profileFilter,
+                                GridBagConstraints(
+                                    1, 0, 1, 1, 1.0, 1.0, GridBagConstraints.BASELINE_TRAILING,
+                                    GridBagConstraints.HORIZONTAL, JBUI.emptyInsets(), 0, 0
+                                )
+                            )
+                        },
+                        BorderLayout.NORTH
+                    )
+                }
+                secondComponent = preview.component
+            },
+            BorderLayout.CENTER
         )
-        northPanel.add(
-            profileFilter,
-            GridBagConstraints(
-                1, 0, 1, 1, 1.0, 1.0, GridBagConstraints.BASELINE_TRAILING,
-                GridBagConstraints.HORIZONTAL, JBUI.emptyInsets(), 0, 0
-            )
-        )
-        treePanel.add(northPanel, BorderLayout.NORTH)
-        return centerPanel
     }
 
     /**
@@ -248,15 +234,14 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
         fillTreeData(null, true)
 
         val renderer = object : TemplateTreeRenderer() {
-            override val filter: String?
-                get() = profileFilter?.filter
+            override val filter
+                get() = profileFilter.filter
         }
 
         tree = object : CheckboxTree(renderer, root) {
             override fun getPreferredScrollableViewportSize(): Dimension {
-                var size = super.getPreferredScrollableViewportSize()
-                size = Dimension(size.width + 10, size.height)
-                return size
+                val size = super.getPreferredScrollableViewportSize()
+                return Dimension(size.width + 10, size.height)
             }
 
             override fun onNodeStateChanged(node: CheckedTreeNode) {
@@ -277,7 +262,6 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
             TreeUtil.expandAll(this)
 
             treeExpander = DefaultTreeExpander(this)
-            profileFilter = TemplatesFilterComponent()
         }
 
         return ScrollPaneFactory.createScrollPane(tree).apply {
@@ -293,64 +277,62 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
      */
     private fun createTreeActionsToolbarPanel(target: JComponent): ActionToolbar {
         val actionManager = CommonActionsManager.getInstance()
-        val actions = DefaultActionGroup()
-        actions.add(actionManager.createExpandAllAction(treeExpander, tree))
-        actions.add(actionManager.createCollapseAllAction(treeExpander, tree))
-        actions.add(
-            object : AnAction(
-                message("dialog.generator.unselectAll"),
-                null,
-                AllIcons.Actions.Unselectall
-            ) {
-                override fun update(e: AnActionEvent) {
-                    e.presentation.isEnabled = checked.isNotEmpty()
-                }
+        val actions = DefaultActionGroup().apply {
+            add(actionManager.createExpandAllAction(treeExpander, tree))
+            add(actionManager.createCollapseAllAction(treeExpander, tree))
+            add(
+                object : AnAction(message("dialog.generator.unselectAll"), null, AllIcons.Actions.Unselectall) {
+                    override fun update(e: AnActionEvent) {
+                        e.presentation.isEnabled = checked.isNotEmpty()
+                    }
 
-                override fun actionPerformed(e: AnActionEvent) {
-                    checked.clear()
-                    filterTree(profileFilter!!.textEditor.text)
-                }
-            }
-        )
-        actions.add(
-            object : AnAction(message("dialog.generator.star"), null, STAR) {
-                override fun update(e: AnActionEvent) {
-                    val node = currentNode
-                    val disabled = node == null || Resources.Template.Container.USER == node.container || !node.isLeaf
-                    val unstar = node != null && Resources.Template.Container.STARRED == node.container
-                    val icon = if (disabled) IconLoader.getDisabledIcon(STAR) else if (unstar) IconLoader.getTransparentIcon(STAR) else STAR
-                    val text = message(if (unstar) "dialog.generator.unstar" else "dialog.generator.star")
-                    val presentation = e.presentation
-                    presentation.isEnabled = !disabled
-                    presentation.icon = icon
-                    presentation.text = text
-                }
-
-                override fun actionPerformed(e: AnActionEvent) {
-                    val node = currentNode ?: return
-                    val template = node.template
-                    if (template != null) {
-                        val isStarred = !template.isStarred
-                        template.isStarred = isStarred
-                        refreshTree()
-                        if (isStarred) {
-                            starred.add(template.name)
-                        } else {
-                            starred.remove(template.name)
-                        }
-                        settings.starredTemplates = ArrayList(starred)
+                    override fun actionPerformed(e: AnActionEvent) {
+                        checked.clear()
+                        filterTree(profileFilter.textEditor.text)
                     }
                 }
+            )
+            add(
+                object : AnAction(message("dialog.generator.star"), null, STAR) {
+                    override fun update(e: AnActionEvent) {
+                        val node = currentNode
+                        val disabled = node == null || USER == node.container || !node.isLeaf
+                        val unstar = node != null && STARRED == node.container
+                        e.presentation.apply {
+                            isEnabled = !disabled
+                            icon = when {
+                                disabled -> IconLoader.getDisabledIcon(STAR)
+                                unstar -> IconLoader.getTransparentIcon(STAR)
+                                else -> STAR
+                            }
+                            text = message(if (unstar) "dialog.generator.unstar" else "dialog.generator.star")
+                        }
+                    }
 
-                /**
-                 * Returns current [TemplateTreeNode] node if available.
-                 *
-                 * @return current node
-                 */
-                private val currentNode
-                    get() = tree?.selectionPath?.lastPathComponent as TemplateTreeNode?
-            }
-        )
+                    override fun actionPerformed(e: AnActionEvent) {
+                        currentNode?.template?.let {
+                            val isStarred = !it.isStarred
+                            it.isStarred = isStarred
+                            refreshTree()
+                            if (isStarred) {
+                                starred.add(it.name)
+                            } else {
+                                starred.remove(it.name)
+                            }
+                            settings.starredTemplates = ArrayList(starred)
+                        }
+                    }
+
+                    /**
+                     * Returns current [TemplateTreeNode] node if available.
+                     *
+                     * @return current node
+                     */
+                    private val currentNode
+                        get() = tree.selectionPath?.lastPathComponent as TemplateTreeNode?
+                }
+            )
+        }
 
         return ActionManager.getInstance().createActionToolbar(ActionPlaces.UNKNOWN, actions, true).apply {
             setTargetComponent(target)
@@ -368,9 +350,8 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
         ApplicationManager.getApplication().runWriteAction {
             CommandProcessor.getInstance().runUndoTransparentAction {
                 val content = if (template != null) StringUtil.notNullize(template.content).replace('\r', '\u0000') else ""
-                previewDocument!!.replaceString(0, previewDocument!!.textLength, content)
-                val pairs = getFilterRanges(profileFilter!!.textEditor.text, content)
-                highlightWords(pairs)
+                previewDocument.replaceString(0, previewDocument.textLength, content)
+                highlightWords(getFilterRanges(profileFilter.textEditor.text, content))
             }
         }
     }
@@ -384,19 +365,21 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
     private fun fillTreeData(filter: String?, forceInclude: Boolean) {
         root.removeAllChildren()
         root.isChecked = false
-        for (container in Resources.Template.Container.values()) {
-            val node = TemplateTreeNode(container)
-            node.isChecked = false
-            root.add(node)
+        Container.values().forEach {
+            root.add(
+                TemplateTreeNode(it).apply {
+                    isChecked = false
+                }
+            )
         }
-        val templatesList = Resources.gitignoreTemplates
-        for (template in templatesList) {
-            if (filter != null && filter.isNotEmpty() && !isTemplateAccepted(template, filter)) {
-                continue
+        Resources.gitignoreTemplates.forEach {
+            if (isTemplateAccepted(it, filter)) {
+                getGroupNode(root, it.container).add(
+                    TemplateTreeNode(it).apply {
+                        isChecked = checked.contains(it)
+                    }
+                )
             }
-            val node = TemplateTreeNode(template)
-            node.isChecked = checked.contains(template)
-            getGroupNode(root, template.container).add(node)
         }
         if (filter != null && forceInclude && root.childCount == 0) {
             fillTreeData(filter, false)
@@ -411,18 +394,14 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
      * @param content templates content
      * @return text ranges
      */
-    private fun getFilterRanges(filter: String, content: String): List<Pair<Int, Int>> {
-        var content = content
-        val pairs: MutableList<Pair<Int, Int>> = ArrayList()
-        content = content.toLowerCase()
-        for (word in getWords(filter)) {
-            var index = content.indexOf(word)
+    private fun getFilterRanges(filter: String, content: String) = mutableListOf<Pair<Int, Int>>().apply {
+        getWords(filter).forEach {
+            var index = content.indexOf(it, ignoreCase = true)
             while (index >= 0) {
-                pairs.add(Pair.create(index, index + word.length))
-                index = content.indexOf(word, index + 1)
+                add(Pair.create(index, index + it.length))
+                index = content.indexOf(it, index + 1, ignoreCase = true)
             }
         }
-        return pairs
     }
 
     /**
@@ -432,22 +411,11 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
      * @param filter   templates filter
      * @return template is accepted
      */
-    private fun isTemplateAccepted(template: Resources.Template, filter: String): Boolean {
-        var filter = filter
-        filter = filter.toLowerCase()
-        if (StringUtil.containsIgnoreCase(template.name, filter)) {
-            return true
-        }
-        var nameAccepted = true
-        for (word in getWords(filter)) {
-            if (!StringUtil.containsIgnoreCase(template.name, word)) {
-                nameAccepted = false
-                break
-            }
-        }
-        val ranges = getFilterRanges(filter, StringUtil.notNullize(template.content))
-        return nameAccepted || ranges.isNotEmpty()
-    }
+    private fun isTemplateAccepted(template: Resources.Template, filter: String?) =
+        filter.isNullOrEmpty() ||
+            template.name.contains(filter, true) ||
+            !getWords(filter).any { !template.name.contains(it, true) } ||
+            getFilterRanges(filter, StringUtil.notNullize(template.content)).isNotEmpty()
 
     /**
      * Filters templates tree.
@@ -455,19 +423,17 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
      * @param filter text
      */
     private fun filterTree(filter: String?) {
-        if (tree != null) {
-            fillTreeData(filter, true)
-            reloadModel()
-            TreeUtil.expandAll(tree!!)
-            if (tree!!.selectionPath == null) {
-                TreeUtil.selectFirstNode(tree!!)
-            }
+        fillTreeData(filter, true)
+        reloadModel()
+        TreeUtil.expandAll(tree)
+        if (tree.selectionPath == null) {
+            TreeUtil.selectFirstNode(tree)
         }
     }
 
     /** Refreshes current tree. */
     private fun refreshTree() {
-        filterTree(profileFilter!!.textEditor.text)
+        filterTree(profileFilter.textEditor.text)
     }
 
     /**
@@ -476,11 +442,12 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
      * @param pairs text ranges
      */
     private fun highlightWords(pairs: List<Pair<Int, Int>>) {
-        val attr = TextAttributes()
-        attr.backgroundColor = UIUtil.getTreeSelectionBackground(true)
-        attr.foregroundColor = UIUtil.getTreeSelectionForeground(true)
-        for (pair in pairs) {
-            preview!!.markupModel.addRangeHighlighter(
+        val attr = TextAttributes().apply {
+            backgroundColor = UIUtil.getTreeSelectionBackground(true)
+            foregroundColor = UIUtil.getTreeSelectionForeground(true)
+        }
+        pairs.forEach { pair ->
+            preview.markupModel.addRangeHighlighter(
                 pair.first,
                 pair.second,
                 0,
@@ -492,10 +459,10 @@ class GeneratorDialog(private val project: Project, var file: PsiFile? = null, v
 
     /** Reloads tree model. */
     private fun reloadModel() {
-        (tree!!.model as DefaultTreeModel).reload()
+        (tree.model as DefaultTreeModel).reload()
     }
 
-    private inner class TemplatesFilterComponent : FilterComponent(TEMPLATES_FILTER_HISTORY, 10) {
+    inner class TemplatesFilterComponent : FilterComponent(TEMPLATES_FILTER_HISTORY, 10) {
         override fun filter() {
             filterTree(filter)
         }